tests/wpt/tests/css/css-conditional/container-queries/scroll-state/at-container-overflowing-serialization.html

<!DOCTYPE html>
<title>CSS Container Queries: scroll-state(overflowing) conditionText serialization</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<link rel="help" href="https://drafts.csswg.org/cssom/#serialize-a-css-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<style id="testSheet">
  @container scroll-state(        overflowing:left) { }
  @container scroll-STate(overflowing:    ) { }
  @container scroll-STate(overflowing) { }
  @container  scroll-state(  ( overflowing: LEFT) OR ( OVERFLOWING: BOTTOM  ) ) { }
  @container scroll-state (overflowing: right) { }
</style>
<script>
  setup(() => {
    assert_implements_container_queries();
    assert_equals(testSheet.sheet.cssRules.length, 5);
  });

  const tests = [
      ["scroll-state(overflowing: left)", "Normalize spaces"],
      ["scroll-STate(overflowing:    )", "No value - invalid, serializes as <general-enclosed>"],
      ["scroll-state(overflowing)", "Boolean context"],
      ["scroll-state((overflowing: left) or (overflowing: bottom))", "Logical with 'or'"],
      ["scroll-state (overflowing: right)", "Not a scroll-state function with space before '('"]
  ].map((e, i) => [testSheet.sheet.cssRules[i], ...e]);

  tests.forEach((t) => {
    test(() => assert_equals(t[0].conditionText, t[1]), t[2]);
  });
</script>
